requirements.txt
appdirs
attrs
mttkinter
raven
#
# dev only
sphinx
semantic_version
wheel
black
pytest
coverage
pytest-cov
PyQt5
pyautogui
